Change location of performance test to the project folder and out of the main repository
