+++ TestProject/Usd-Performance/Assets/Artifacts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e6375da77b9195e4cb6b6cc7277c6830
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
