+++ TestProject/Usd-Performance/Assets/Artifacts/44e94a84-9ecb-4c1d-8c78-f6a92906e4e4.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 5012970d63036294085e7c5c4aea826a
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
