+++ TestProject/Usd-Performance/Assets/Artifacts/550db173-2b24-4237-aecb-d05d6c83de7b.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 6835039162df7ca4eabd9b6503401464
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
