+++ TestProject/Usd-Performance/Assets/Artifacts/70809873-e545-4b67-9670-94000a43651f.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 65119642d223a57499dc3df426dc1b16
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
