+++ TestProject/Usd-Performance/Assets/Artifacts/b864caf4-d8aa-441f-8ab6-113d42990f70.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: bf67ce5d31205cb49a504a1e09f4c0c2
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
