+++ TestProject/Usd-Performance/Assets/Artifacts/e7254af6-0aa4-469f-94f4-d20f0315d6d1.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d436d1b9a0820d44da6d315148bbcf4a
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
