+++ TestProject/Usd-Performance/Assets/Artifacts/f08edebb-283c-4d90-81fe-fbe3af657b62.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 7488715688f0b584cac45e79673ed7af
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
