+++ TestProject/Usd-Performance/Assets/Performance.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 624667d1ff65b354dbab0b67db4ea4ab
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
