+++ TestProject/Usd-Performance/Assets/Performance/BaseUsdPerformanceTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fd9eac63ffef6574792a5611cb1187b6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
