+++ TestProject/Usd-Performance/Assets/Performance/Data.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7dea1eb830c53244f82a091816b4ff73
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
