+++ TestProject/Usd-Performance/Assets/Performance/Data/PointInstancer.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: aafa1a2a72eee0d438a095d600f5464c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
