+++ TestProject/Usd-Performance/Assets/Performance/Data/PointInstancer/instanced_cubes_1000.usda.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 9ca154d52e7e03e4c989b66e33e08212
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
