+++ TestProject/Usd-Performance/Assets/Performance/PerformanceBaseFixture.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4aaa5bac57b21eb4bbc46f4394de2701
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
