+++ TestProject/Usd-Performance/Assets/Performance/Unity.Formats.USD.Tests.Performance.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 84db87f57b3897149bdd2a0bd3d2f966
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
