+++ TestProject/Usd-Performance/ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 2019.4.26f1
-m_EditorVersionWithRevision: 2019.4.26f1 (e0392c6b2363)
+m_EditorVersion: 2020.3.46f1
+m_EditorVersionWithRevision: 2020.3.46f1 (18bc01a066b4)
