--- package/com.unity.formats.usd/Tests/Performance/Data/PointInstancer.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: aafa1a2a72eee0d438a095d600f5464c
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
