--- package/com.unity.formats.usd/Tests/Performance/Data/PointInstancer/instanced_cubes_10.usda.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: ddba66229ec63364d8ce36112afcdbec
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
