--- package/com.unity.formats.usd/Tests/Performance/Data/PointInstancer/instanced_cubes_100.usda
@@ -1,32 +1,0 @@
-#usda 1.0
-(
-    endTimeCode = 1
-    framesPerSecond = 24
-    metersPerUnit = 1
-    startTimeCode = 1
-    timeCodesPerSecond = 24
-    upAxis = "Y"
-)
-
-def PointInstancer "instancer"
-{
-    int64[] invisibleIds = []
-    quath[] orientations = [(1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0), (1, 0, 0, 0)]
-    point3f[] positions = [(-0.16644895, 0.3755865, 0.3704425), (0.33355105, -0.12441349, -0.12955749), (-0.41644895, 0.12558651, -0.3795575), (0.08355105, -0.3744135, 0.12044251), (-0.29144895, -0.24941349, -0.0045574903), (0.20855105, 0.2505865, 0.4954425), (-0.04144895, -0.4994135, 0.24544251), (0.45855105, 0.0005865097, -0.2545575), (-0.47894895, -0.4369135, 0.4329425), (0.02105105, 0.06308651, -0.06705749), (-0.22894895, -0.18691349, -0.3170575), (0.27105105, 0.3130865, 0.18294251), (-0.10394895, 0.18808651, -0.19205749), (0.39605105, -0.3119135, 0.3079425), (-0.35394895, 0.4380865, 0.05794251), (0.14605105, -0.06191349, -0.4420575), (-0.32269895, 0.09433651, 0.15169251), (0.17730105, -0.4056635, -0.3483075), (-0.07269895, 0.3443365, -0.09830749), (0.42730105, -0.15566349, 0.4016925), (-0.19769895, -0.2806635, -0.4733075), (0.30230105, 0.21933651, 0.02669251), (-0.44769895, -0.03066349, 0.2766925), (0.05230105, 0.4693365, -0.22330749), (-0.010198951, -0.09316349, 0.08919251), (0.48980105, 0.4068365, -0.4108075), (-0.26019895, -0.3431635, -0.16080749), (0.23980105, 0.15683651, 0.3391925), (-0.38519895, 0.2818365, -0.2858075), (0.11480105, -0.21816349, 0.21419251), (-0.13519895, 0.03183651, 0.4641925), (0.36480105, -0.4681635, -0.03580749), (-0.43207395, -0.17128849, -0.14518249), (0.06792605, 0.3287115, 0.3548175), (-0.18207395, -0.4212885, 0.10481751), (0.31792605, 0.07871151, -0.3951825), (-0.05707395, 0.4537115, 0.4798175), (0.44292605, -0.04628849, -0.02018249), (-0.30707395, 0.20371151, -0.2701825), (0.19292605, -0.2962885, 0.22981751), (-0.24457395, 0.14121151, -0.08268249), (0.25542605, -0.3587885, 0.4173175), (-0.49457395, 0.3912115, 0.16731751), (0.005426049, -0.10878849, -0.3326825), (-0.36957395, -0.4837885, 0.2923175), (0.13042605, 0.01621151, -0.20768249), (-0.11957395, -0.23378849, -0.4576825), (0.38042605, 0.2662115, 0.04231751), (-0.08832395, -0.3275385, -0.3639325), (0.41167605, 0.17246151, 0.13606751), (-0.33832395, -0.07753849, 0.3860675), (0.16167605, 0.4224615, -0.11393249), (-0.46332395, 0.04746151, 0.01106751), (0.03667605, -0.4525385, -0.4889325), (-0.21332395, 0.2974615, -0.23893249), (0.28667605, -0.20253849, 0.2610675), (-0.27582395, 0.3599615, -0.4264325), (0.22417605, -0.14003849, 0.07356751), (-0.02582395, 0.10996151, 0.3235675), (0.47417605, -0.3900385, -0.17643249), (-0.15082395, -0.01503849, 0.19856751), (0.34917605, 0.4849615, -0.3014325), (-0.40082395, -0.2650385, -0.05143249), (0.09917605, 0.23496151, 0.4485675), (-0.29926145, -0.304101, 0.06575501), (0.20073855, 0.19589901, -0.434245), (-0.04926145, -0.05410099, -0.18424499), (0.45073855, 0.445899, 0.315755), (-0.17426145, 0.07089901, -0.309245), (0.32573855, -0.429101, 0.19075501), (-0.42426145, 0.320899, 0.440755), (0.07573855, -0.17910099, -0.05924499), (-0.11176145, 0.258399, 0.12825501), (0.38823855, -0.24160099, -0.371745), (-0.36176145, 0.00839901, -0.12174499), (0.13823855, -0.491601, 0.378255), (-0.48676145, -0.11660099, -0.496745), (0.013238549, 0.383399, 0.0032550097), (-0.23676145, -0.366601, 0.253255), (0.26323855, 0.13339901, -0.24674499), (-0.20551145, -0.21035099, 0.409505), (0.29448855, 0.289649, -0.09049499), (-0.45551145, -0.460351, -0.340495), (0.04448855, 0.03964901, 0.15950501), (-0.33051145, 0.414649, -0.21549499), (0.16948855, -0.08535099, 0.284505), (-0.08051145, 0.16464901, 0.03450501), (0.41948855, -0.335351, -0.465495), (-0.39301145, 0.22714901, 0.347005), (0.10698855, -0.272851, -0.15299499), (-0.14301145, 0.477149, -0.402995), (0.35698855, -0.02285099, 0.09700501), (-0.01801145, -0.397851, -0.02799499), (0.48198855, 0.10214901, 0.472005), (-0.26801145, -0.14785099, 0.22200501), (0.23198855, 0.352149, -0.277995), (-0.03363645, 0.02402401, -0.41862), (0.46636355, -0.475976, 0.08138001), (-0.28363645, 0.274024, 0.33138), (0.21636355, -0.22597599, -0.16861999)]
-    int[] protoIndices = [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]
-    rel prototypes = </instancer/Prototypes/Root>
-
-    def Scope "Prototypes"
-    {
-        def Xform "Root"
-        {
-            def Cube "cube1"
-            {
-                double size = 0.10000000149011612
-                matrix4d xformOp:transform = ( (1, 0, 0, 0), (0, 1, 0, 0), (0, 0, 1, 0), (0, 0, 0, 1) )
-                uniform token[] xformOpOrder = ["xformOp:transform"]
-            }
-        }
-    }
-}
-
