--- package/com.unity.formats.usd/Tests/Performance/Data/PointInstancer/instanced_cubes_100.usda.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 567d2c32dba4a364bb1fb80b134dc92d
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
