--- package/com.unity.formats.usd/Tests/Performance/Data/PointInstancer/instanced_cubes_1000.usda.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 9ca154d52e7e03e4c989b66e33e08212
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
