--- package/com.unity.formats.usd/Tests/Performance/Data/PointInstancer/instanced_cubes_10000.usda.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: b4f2b0c08e1cc0d43aa738154881fcfc
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
