--- package/com.unity.formats.usd/Tests/Performance/PointInstancerPerformanceTests.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 393dc8fe2fdf90149b716736a75f70ac
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
