--- package/com.unity.formats.usd/Tests/Performance/Unity.Formats.USD.Tests.Performance.asmdef
@@ -1,30 +1,0 @@
-{
-    "name": "Unity.Formats.USD.Tests.Performance",
-    "rootNamespace": "",
-    "references": [
-        "GUID:ff4f2b49d51b1be44b50586a72af45e7",
-        "GUID:0acc523941302664db1f4e527237feb3",
-        "GUID:27619889b8ba8c24980f49ee34dbb44a",
-        "GUID:88d59bdd77f09d744b8708dbf107c107",
-        "GUID:7263a266704d56b44818388e07b5f09b",
-        "GUID:c0dd0d10738d4ad4a9de57c559d0ca1b",
-        "GUID:e50fcb8f70e44487fae2be8974997e9c",
-        "GUID:e31ceb1e0618b432f826d902eeb8e779",
-        "GUID:fbb89535e17e84eae8079c2158650d01"
-    ],
-    "includePlatforms": [
-        "Editor"
-    ],
-    "excludePlatforms": [],
-    "allowUnsafeCode": true,
-    "overrideReferences": true,
-    "precompiledReferences": [
-        "nunit.framework.dll"
-    ],
-    "autoReferenced": false,
-    "defineConstraints": [
-        "UNITY_INCLUDE_TESTS"
-    ],
-    "versionDefines": [],
-    "noEngineReferences": false
-}
